+++ src/kili/entrypoints/mutations/label/__init__.py
@@ -1,6 +1,5 @@
 """Label mutations."""
 
-import warnings
 from json import dumps
 from typing import Dict, List, Literal, Optional, cast
 
@@ -17,9 +16,6 @@ from kili.entrypoints.mutations.label.queries import (
     GQL_APPEND_TO_LABELS,
     GQL_CREATE_HONEYPOT,
 )
-from kili.presentation.client.helpers.common_validators import (
-    assert_all_arrays_have_same_size,
-)
 from kili.use_cases.asset.utils import AssetUseCasesUtils
 from kili.utils.logcontext import for_all_methods, log_call
 
@@ -27,88 +23,6 @@ from kili.utils.logcontext import for_all_methods, log_call
 @for_all_methods(log_call, exclude=["__init__"])
 class MutationsLabel(BaseOperationEntrypointMixin):
     """Set of Label mutations."""
-
-    # pylint: disable=too-many-arguments
-    @typechecked
-    def create_predictions(
-        self,
-        project_id: str,
-        external_id_array: Optional[List[str]] = None,
-        model_name_array: Optional[List[str]] = None,
-        json_response_array: Optional[List[dict]] = None,
-        model_name: Optional[str] = None,
-        asset_id_array: Optional[List[str]] = None,
-        disable_tqdm: Optional[bool] = None,
-        overwrite: bool = False,
-    ) -> Dict[Literal["id"], str]:
-        # pylint: disable=line-too-long
-        """Create predictions for specific assets.
-
-        Args:
-            project_id: Identifier of the project.
-            external_id_array: The external IDs of the assets for which we want to add predictions.
-            model_name_array: Deprecated, use `model_name` instead.
-            json_response_array: The predictions are given here. For examples,
-                see [the recipe](https://docs.kili-technology.com/recipes/importing-labels-and-predictions).
-            model_name: The name of the model that generated the predictions
-            asset_id_array: The internal IDs of the assets for which we want to add predictions.
-            disable_tqdm: Disable tqdm progress bar.
-            overwrite: if True, it will overwrite existing predictions of
-                the same model name on the targeted assets.
-
-        Returns:
-            A dictionary with the project `id`.
-
-        !!! example "Recipe"
-            For more detailed examples on how to create predictions, see [the recipe](https://docs.kili-technology.com/recipes/importing-labels-and-predictions).
-
-        !!! warning "model name"
-            The use of `model_name_array` is deprecated. Creating predictions from different
-            models is not supported anymore. Please use `model_name` argument instead to
-            provide the predictions model name.
-        """
-        if json_response_array is None or len(json_response_array) == 0:
-            raise ValueError(
-                "json_response_array is empty, you must provide at least one prediction to upload"
-            )
-        assert_all_arrays_have_same_size(
-            [external_id_array, json_response_array, model_name_array, asset_id_array]
-        )
-        if model_name is None:
-            if model_name_array is None:
-                raise ValueError("You must provide a model name with the model_name argument ")
-            if len(set(model_name_array)) > 1:
-                raise ValueError(
-                    "Creating predictions from different models is not supported anymore. Separate"
-                    " your calls by models."
-                )
-            warnings.warn(
-                "The use of `model_name_array` is deprecated. Creating predictions from"
-                " different models is not supported anymore. Please use `model_name` argument"
-                " instead to provide the predictions model name.",
-                DeprecationWarning,
-                stacklevel=1,
-            )
-            model_name = model_name_array[0]
-
-        labels = [
-            {
-                "asset_id": asset_id,
-                "asset_external_id": asset_external_id,
-                "json_response": json_response,
-            }
-            for (asset_id, asset_external_id, json_response) in list(
-                zip(
-                    asset_id_array or [None] * len(json_response_array),
-                    external_id_array or [None] * len(json_response_array),
-                    json_response_array,
-                )
-            )
-        ]
-        import_labels_from_dict(
-            self, project_id, labels, "PREDICTION", overwrite, model_name, disable_tqdm
-        )
-        return {"id": project_id}
 
     @deprecate(
         msg=(
