+++ src/kili/services/label_import/__init__.py
@@ -3,7 +3,6 @@
 from pathlib import Path
 from typing import List, Optional, Type, cast
 
-from kili.domain.label import LabelType
 from kili.domain.project import ProjectId
 from kili.exceptions import NotFound
 from kili.services.helpers import is_target_job_in_json_interface
